requirements.txt
PyYAML~=3.13
pip~=20.2
configparser~=3.5.0
wheel~=0.34.2
cryptography~=2.3
lxml~=4.2.3
tornado~=5.1.1
typing~=3.6.4
keyring~=18.0.1
attrs~=19.3.0
six~=1.15.0
pyasn1~=0.4.3
future~=0.18.2
bcrypt~=3.1.4
appdirs~=1.4.4
PyGObject~=3.36.0
click~=6.7
Pygments~=2.3.1
aiohttp~=2.3.9
SQLAlchemy~=1.2.10
idna~=2.7
pycparser~=2.18
setuptools~=40.0.0
docutils~=0.15.2
twisted
service-identity